Calculate the Canadian AQHI from hourly PM2.5, O3, and NO2 concentrations

The Canadian Air Quality Health Index (AQHI) is a measure of the health risk associated with exposure to ambient air pollution. It combines the impacts of fine particulate matter (PM2.5), ozone (O3), and nitrogen dioxide (NO2) on a scale from 1-10 (+).

AQHI formula: 10⁄10.4 × 100 × (e0.000537 * O3 - 1) + (e0.000871 * NO2 - 1) + (e0.000487 * PM2.5 - 1)

The AQHI was originally published by Stieb et al. (2008): doi:10.3155/1047-3289.58.3.435 , and is used by all Canadian provinces and territories except for Québec, which uses the AQI instead. The AQHI is calculated for each hour using 3-hour rolling mean concentrations of PM2.5, O3, and NO2. However, the AQHI is overriden by the AQHI+ for a given hour if that AQHI+ level is higher. The AQHI+ is a modification of the AQHI that only uses PM2.5 and is calculated using hourly means (see AQHI_plus).

Usage

AQHI(
  dates,
  pm25_1hr_ugm3,
  o3_1hr_ppb = NA_real_,
  no2_1hr_ppb = NA_real_,
  allow_aqhi_plus_override = TRUE,
  detailed = TRUE,
  language = "en",
  quiet = FALSE
)

Arguments

dates

A POSIXct (date-time) vector corresponding to hourly pollutant observations. Date gaps will be filled with NA's automatically as needed when calculating 3-hour averages.

pm25_1hr_ugm3

A numeric vector of hourly mean PM2.5 concentrations (units = μg m-3). Must be a single value or the same length as dates or a single value (which will be repeated).

o3_1hr_ppb

(Optional). A numeric vector of hourly mean ozone (O3) concentrations (units = ppb). Must be a single value or the same length as dates or a single value (which will be repeated). If all are NA (the default) AQHI won't be calculated and instead AQHI+ (see AQHI_plus) will be used.

no2_1hr_ppb

(Optional). A numeric vector of hourly mean nitrogen dioxide (NO2) concentrations (units = ppb). Must be a single value or the same length as dates or a single value (which will be repeated). If all are NA (the default) AQHI won't be calculated and instead AQHI+ (see AQHI_plus) will used.

allow_aqhi_plus_override

(Optional). A single logical value indicating if the AQHI+ should be allowed to override the AQHI if it exceeds the AQHI for a particular hour. Default is TRUE.

detailed

(Optional). A single logical value indicating if a tibble with levels, risk categories, health messages, etc should be returned. If FALSE only the levels will be returned. Default is TRUE.

language

(Optional). A single character value indicating the language to use for risk levels and health messaging. Must be either "en" (English) or "fr" (French). Not case sensitive. Default is "en".

quiet

(Optional). A single logical (TRUE/FALSE) value indicating if non-critical warnings/messages should be hidden. Default is FALSE.

Value

If detailed = TRUE:

A tibble (data.frame) with a detailed summary of the AQHI, risk levels, health messages, etc and the same number of rows as length(dates)

If detailed = FALSE:

a factor vector of AQHI levels with length(dates) elements

References

Stieb et al. (2008): doi:10.3155/1047-3289.58.3.435

Environment and Climate Change Canada: https://www.canada.ca/en/environment-climate-change/services/air-quality-health-index/about.html

Examples

obs <- example_obs[example_obs$site_id == 1, ]

# Calculate the AQHI
obs$date_utc |>
  AQHI(
    pm25_1hr_ugm3 = obs$pm25_1hr,
    o3_1hr_ppb = obs$o3_1hr,
    no2_1hr_ppb = obs$no2_1hr
  )
#> # A tibble: 8,571 × 18
#>    date                pm25_1hr_ugm3 o3_1hr_ppb no2_1hr_ppb pm25_3hr_ugm3
#>    <dttm>                      <dbl>      <dbl>       <dbl>         <dbl>
#>  1 2018-01-01 09:00:00           9.2        0.7        18.1          NA  
#>  2 2018-01-01 10:00:00          12.1        0.3        17.5          NA  
#>  3 2018-01-01 11:00:00          10.7        0.2        17.8          10.7
#>  4 2018-01-01 12:00:00           9.3        0.2        18.1          10.7
#>  5 2018-01-01 13:00:00           8          1.1        16.6           9.3
#>  6 2018-01-01 14:00:00           6.6        1.4        15.6           8  
#>  7 2018-01-01 15:00:00           6          1.9        13.8           6.9
#>  8 2018-01-01 16:00:00           6.1        0.3        16.9           6.2
#>  9 2018-01-01 17:00:00           7.1        1          18.1           6.4
#> 10 2018-01-01 18:00:00           7.9        2.5        16.9           7  
#> # ℹ 8,561 more rows
#> # ℹ 13 more variables: o3_3hr_ppb <dbl>, no2_3hr_ppb <dbl>, level <fct>,
#> #   AQHI <fct>, AQHI_plus <fct>, AQHI_plus_exceeds_AQHI <lgl>,
#> #   AQHI_pm25_ratio <dbl>, AQHI_o3_ratio <dbl>, AQHI_no2_ratio <dbl>,
#> #   colour <chr>, risk <fct>, high_risk_pop_message <chr>,
#> #   general_pop_message <chr>

# Return just the AQHI levels
obs$date_utc |>
  AQHI(
    pm25_1hr_ugm3 = obs$pm25_1hr,
    o3_1hr_ppb = obs$o3_1hr,
    no2_1hr_ppb = obs$no2_1hr,
    detailed = FALSE
  )
#>    [1] 1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  2  2  2  2  2  2 
#>   [25] 2  2  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#>   [49] 1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  1  1  1  1  1 
#>   [73] 1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  1  1  1  1  1  1 
#>   [97] 1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#>  [121] 2  2  2  2  2  2  2  2  2  2  2  3  3  3  2  2  2  2  2  2  2  2  2  2 
#>  [145] 2  2  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#>  [169] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#>  [193] 2  2  2  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1 
#>  [217] 2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  2  2  2  2  2  2  1  1 
#>  [241] 1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  2  2 
#>  [265] 3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  2  2  2  2  2 
#>  [289] 2  2  2  2  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1 
#>  [313] 1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1 
#>  [337] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  1  1  1  1  1  1 
#>  [361] 1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1 
#>  [385] 1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#>  [409] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#>  [433] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#>  [457] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#>  [481] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  2  2  2  2  2 
#>  [505] 2  2  2  2  2  2  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  1  1  1 
#>  [529] 1  2  2  2  2  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#>  [553] 2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  2  2  2  2  2  2  2  2  2 
#>  [577] 2  2  2  2  2  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#>  [601] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  2  2  2  2  2  2 
#>  [625] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#>  [649] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  2  2  2  2  1  3 
#>  [673] 3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  1  2  1  1  1  2 
#>  [697] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#>  [721] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  2  2  2  2  2 
#>  [745] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  2  2 
#>  [769] 2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [793] 1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  2  2  2 
#>  [817] 2  2  2  2  3  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1 
#>  [841] 1  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [865] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  1  1  1  2  2  2 
#>  [889] 2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  2  2  2  2  2 
#>  [913] 2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  2  2 
#>  [937] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1 
#>  [961] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3 
#>  [985] 2  3  3  3  3  3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2 
#> [1009] 2  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1033] 2  2  3  3  2  2  2  3  3  3  3  3  2  2  2  2  2  1  1  1  1  1  2  2 
#> [1057] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1081] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1105] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1129] 2  2  2  2  2  2  2  2  3  3  3  2  2  2  2  2  2  2  3  3  3  3  2  2 
#> [1153] 2  2  3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  3  3  2  2  2  2 
#> [1177] 2  2  2  2  3  3  3  3  3  3  3  3  2  2  2  2  2  2  3  3  3  3  3  3 
#> [1201] 2  2  2  2  2  2  3  3  3  2  2  3  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1225] 2  3  3  3  3  3  3  3  3  3  3  3  2  2  2  2  2  3  3  3  3  2  2  2 
#> [1249] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1273] 2  2  2  2  2  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1297] 2  2  2  2  2  2  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1321] 3  3  3  3  3  4  3  3  3  3  3  2  2  2  2  2  2  2  2  2  3  3  3  2 
#> [1345] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3 
#> [1369] 2  2  2  2  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3 
#> [1393] 3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1417] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  2  2 
#> [1441] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  2  2  2  2 
#> [1465] 2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  2  4  3  3 
#> [1489] 3  3  3  3  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  2  2 
#> [1513] 2  2  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1537] 2  3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1561] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  2  2  3  3  3  3  3  3  3 
#> [1585] 3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1609] 3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3 
#> [1633] 3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1657] 2  2  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  4 
#> [1681] 3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3 
#> [1705] 3  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  2  2  2  2 
#> [1729] 2  2  2  2  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3 
#> [1753] 3  3  2  3  3  3  3  3  2  2  2  2  2  2  1  1  2  2  2  2  2  2  2  2 
#> [1777] 2  2  2  3  3  3  3  3  3  3  2  2  2  1  1  1  1  1  1  2  2  2  2  2 
#> [1801] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  2  2  2  3  3 
#> [1825] 3  3  3  3  3  3  2  2  2  2  2  2  2  1  1  1  1  2  2  2  2  2  2  2 
#> [1849] 2  2  2  2  2  2  3  3  3  3  2  2  1  1  1  1  1  1  2  2  2  2  2  2 
#> [1873] 2  2  2  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3 
#> [1897] 3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1921] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1945] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1969] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [1993] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [2017] 2  2  2  2  2  2  2  2  1  1  1  1  2  1  1  1  2  2  2  2  3  3  3  2 
#> [2041] 2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  2  2  2  2  2  2  2 
#> [2065] 2  2  2  3  2  2  2  2  2  2  1  1  1  1  1  1  2  2  2  2  2  2  2  2 
#> [2089] 2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2 
#> [2113] 2  2  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3 
#> [2137] 3  3  3  3  3  3  3  2  2  3  2  2  2  2  2  2  2  2  2  2  2  2  3  3 
#> [2161] 3  3  2  3  3  3  3  3  3  3  3  3  3  2  2  2  1  3  3  2  2  3  3  2 
#> [2185] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  2  2  2 
#> [2209] 2  2  2  1  1  1  1  2  2  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [2233] 2  2  2  2  2  2  2  2  1  2  2  1  1  1  1  1  1  1  1  1  1  2  2  2 
#> [2257] 2  2  2  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [2281] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [2305] 2  2  2  2  1  2  2  2  3  2  2  2  1  1  1  1  1  1  1  2  2  2  2  3 
#> [2329] 3  3  3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [2353] 2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  2  2  2  2  3  3  3  3 
#> [2377] 3  3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  3  3  3  2  2 
#> [2401] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3 
#> [2425] 3  3  3  3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2 
#> [2449] 3  3  3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [2473] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [2497] 2  2  2  2  3  3  3  3  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  2 
#> [2521] 2  2  2  2  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  1  1  3  1  3 
#> [2545] 3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3 
#> [2569] 3  3  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  3  3  3  3 
#> [2593] 3  3  3  3  3  3  2  2  2  2  2  2  1  1  1  2  2  2  3  3  3  3  3  3 
#> [2617] 3  3  3  2  2  2  2  2  1  1  1  2  2  2  2  2  2  2  3  3  3  3  3  3 
#> [2641] 3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3 
#> [2665] 3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  3  3  4  3  3  3  3 
#> [2689] 3  3  4  4  3  3  3  3  3  3  2  2  2  3  3  2  3  3  3  4  4  4  4  4 
#> [2713] 4  4  4  4  3  3  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  3  3 
#> [2737] 3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [2761] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [2785] 2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2 
#> [2809] 2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  2  2  2  3  3  3  3  3  3 
#> [2833] 3  3  2  2  2  2  2  2  2  1  1  1  1  1  2  2  2  3  3  3  3  3  3  3 
#> [2857] 3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  3 
#> [2881] 3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3 
#> [2905] 3  3  3  3  2  2  2  2  1  1  1  1  1  1  2  2  2  2  2  2  2  3  3  3 
#> [2929] 3  3  3  2  2  2  2  2  2  1  1  1  1  2  2  3  3  3  3  3  3  3  3  3 
#> [2953] 3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3 
#> [2977] 3  3  3  2  2  2  2  2  2  1  1  1  1  2  2  2  3  3  3  3  3  3  3  3 
#> [3001] 3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  2  2  3  3  3 
#> [3025] 3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  3  3  3 
#> [3049] 3  3  2  2  2  2  1  1  1  1  1  1  1  1  2  2  2  3  3  3  3  3  3  3 
#> [3073] 3  3  2  2  2  2  2  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  3 
#> [3097] 3  3  3  2  2  2  2  2  2  1  1  1  1  1  2  2  3  3  3  3  3  3  4  4 
#> [3121] 4  4  3  3  2  2  2  2  2  2  2  2  2  2  2  3  3  4  4  4  4  4  4  4 
#> [3145] 4  4  4  3  3  3  2  2  2  2  2  2  2  3  3  4  4  3  3  3  4  4  4  4 
#> [3169] 4  4  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3 
#> [3193] 3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [3217] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3 
#> [3241] 3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3 
#> [3265] 3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [3289] 2  3  2  2  2  2  2  2  4  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3 
#> [3313] 3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  3 
#> [3337] 3  3  3  3  2  2  2  2  2  2  2  2  2  2  3  3  4  4  4  4  4  4  4  4 
#> [3361] 4  4  3  3  2  2  2  2  2  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2 
#> [3385] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  3 
#> [3409] 3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3 
#> [3433] 3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  2  2  3 
#> [3457] 3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3 
#> [3481] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3 
#> [3505] 3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3 
#> [3529] 3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [3553] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [3577] 2  2  2  2  2  2  2  2  2  2  1  1  1  1  2  2  2  2  3  3  3  3  3  3 
#> [3601] 3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [3625] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  2  2  2  2 
#> [3649] 2  2  2  2  2  2  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2 
#> [3673] 2  2  2  2  2  2  1  1  1  1  1  1  2  2  2  2  2  3  3  3  3  3  3  3 
#> [3697] 3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3 
#> [3721] 3  2  2  2  2  2  2  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [3745] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [3769] 2  2  2  2  2  2  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2 
#> [3793] 2  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2 
#> [3817] 2  2  2  1  1  1  1  1  1  1  1  2  2  1  3  2  2  2  2  2  2  2  2  1 
#> [3841] 1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1 
#> [3865] 1  1  1  1  1  1  1  1  1  2  1  1  1  1  2  2  2  2  2  2  2  1  1  1 
#> [3889] 1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [3913] 1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  3  3  2  2  2 
#> [3937] 2  1  1  1  1  1  1  1  1  2  3  3  4  3  3  3  3  3  3  3  3  3  3  2 
#> [3961] 2  1  1  1  1  1  1  2  2  2  3  3  4  4  4  4  4  3  3  3  3  3  3  2 
#> [3985] 2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  3  3  3  2  2  2 
#> [4009] 2  2  2  2  2  2  2  2  2  2  3  3  3  3  4  4  4  4  4  4  4  4  3  3 
#> [4033] 2  2  2  2  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [4057] 2  2  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  1  1  2  2 
#> [4081] 2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  2  2  2  2  2  2  3  3 
#> [4105] 3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [4129] 2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  2  2  1  2  2  2 
#> [4153] 2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  2  2  2  2  2  2  2  2  2 
#> [4177] 2  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [4201] 2  2  2  2  2  2  2  2  1  1  1  1  1  2  2  2  1  1  2  2  2  2  2  2 
#> [4225] 2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4249] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2 
#> [4273] 2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2 
#> [4297] 2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2 
#> [4321] 2  2  2  3  3  2  2  2  2  2  1  1  1  1  1  1  2  2  2  3  3  3  3  3 
#> [4345] 3  4  4  3  2  2  2  2  1  1  1  1  1  1  2  1  3  3  3  2  2  3  3  3 
#> [4369] 2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  2  2  2  2  2 
#> [4393] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  1 
#> [4417] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  3  2 
#> [4441] 2  2  2  2  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2 
#> [4465] 2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  2  1  2  2  2  2  2  2  1 
#> [4489] 1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2 
#> [4513] 2  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  3  3  2  2  2  2  2 
#> [4537] 2  2  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [4561] 2  2  2  2  2  2  2  1  1  1  2  2  2  2  2  2  2  2  2  3  3  3  3  3 
#> [4585] 2  2  2  2  2  2  1  1  1  1  2  2  2  3  3  3  3  3  3  2  2  2  2  2 
#> [4609] 2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  3  3  3  2  2  2 
#> [4633] 2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  3  3  3  3  3  2  2 
#> [4657] 2  2  2  2  2  1  1  1  2  2  1  1  2  2  2  2  2  2  2  2  2  2  2  2 
#> [4681] 1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [4705] 2  2  2  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [4729] 2  2  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  6  3  3  2 
#> [4753] 2  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  3  4  3  3  2  2 
#> [4777] 2  2  1  1  1  1  2  2  2  2  3  3  3  3  3  3  3  3  3  3  3  2  2  2 
#> [4801] 2  2  2  2  2  2  2  2  3  3  3  3  3  3  3  3  4  4  3  3  3  3  3  3 
#> [4825] 2  2  2  2  2  2  2  3  3  4  4  4  4  4  4  4  3  3  3  3  3  3  3  2 
#> [4849] 2  2  2  2  2  2  2  2  2  3  1  4  4  4  4  4  4  4  3  3  3  3  3  2 
#> [4873] 2  2  2  3  3  3  3  3  4  4  4  4  4  4  4  4  3  3  3  3  3  3  3  3 
#> [4897] 3  3  3  2  2  2  3  3  3  3  4  4  4  4  4  4  3  3  3  3  3  3  2  2 
#> [4921] 2  2  2  2  3  3  4  4  4  5  5  5  5  5  5  5  5  4  4  3  3  3  3  3 
#> [4945] 3  3  3  3  3  4  2  3  3  4  4  4  3  3  3  3  3  2  2  2  2  2  2  2 
#> [4969] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1 
#> [4993] 1  1  1  1  1  1  1  1  1  1  2  2  2  1  1  1  1  1  1  1  1  1  1  1 
#> [5017] 1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1 
#> [5041] 1  1  1  1  2  2  2  2  2  3  3  3  3  3  3  3  2  2  2  2  2  1  1  1 
#> [5065] 1  1  2  2  3  3  3  3  3  4  4  4  4  4  4  3  3  3  2  2  2  2  2  2 
#> [5089] 2  2  2  2  3  3  4  4  4  4  4  4  4  4  4  4  3  3  2  2  3  2  2  2 
#> [5113] 2  2  2  3  3  4  4  4  4  4  4  4  4  4  4  3  3  3  3  3  3  2  2  2 
#> [5137] 2  3  3  3  4  4  5  5  5  5  5  5  5  5  5  4  3  3  3  3  3  2  2  2 
#> [5161] 2  2  2  2  3  3  3  3  3  3  4  4  4  4  4  4  3  3  3  3  3  3  3  3 
#> [5185] 3  3  3  4  5  5  5  4  4  4  4  4  4  4  3  3  2  2  2  1  1  1  1  1 
#> [5209] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5233] 1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3 
#> [5257] 3  4  4  5  5  5  5  5  5  6  4  4  4  4  4  8  9  8  7  7  7  7  7  7 
#> [5281] 7  8  8  8  8  8  7  7  7  7  8  8  8  7  7  7  7  7  6  6  6  5  5  6 
#> [5305] 6  7  6  5  6  6  5  5  5  5  4  4  5  4  4  3  3  3  3  3  3  3  3  3 
#> [5329] 3  3  3  3  3  3  3  4  4  4  4  4  4  4  3  3  3  3  3  3  3  3  3  3 
#> [5353] 3  2  2  2  3  3  3  3  3  3  4  3  3  3  3  3  3  2  2  2  2  2  2  2 
#> [5377] 2  2  2  3  3  3  3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2 
#> [5401] 2  2  3  5  3  4  3  4  6  6  6  6  5  5  6  6  5  4  5  7  8  8  9  + 
#> [5425] +  +  +  +  +  10 10 9  9  9  8  8  8  7  6  6  6  5  4  4  4  3  3  3 
#> [5449] 4  4  4  6  9  10 10 9  8  8  8  9  8  7  6  6  6  6  7  6  6  5  6  6 
#> [5473] 6  6  6  9  10 10 +  +  +  +  +  +  +  +  +  +  +  +  +  9  6  4  3  4 
#> [5497] 5  5  6  5  4  4  3  3  3  3  3  3  3  2  2  2  2  3  3  3  3  4  4  3 
#> [5521] 3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3 
#> [5545] 3  3  3  4  4  4  4  3  4  4  4  4  4  5  5  4  4  5  4  4  4  4  4  3 
#> [5569] 3  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  2  1  1  1  1  1  1  1 
#> [5593] 1  1  2  2  2  3  3  3  3  3  2  2  2  2  2  2  2  2  1  1  1  1  1  2 
#> [5617] 2  2  2  2  3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  3 
#> [5641] 3  3  3  3  2  3  3  3  3  3  3  3  3  2  2  2  1  1  1  1  1  1  1  1 
#> [5665] 1  1  1  2  2  2  2  2  3  3  2  2  2  2  2  2  1  1  1  1  1  1  1  1 
#> [5689] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  2  2  2  2  1  1  1  1  1 
#> [5713] 1  2  2  2  2  2  3  3  3  3  3  3  3  2  2  2  2  2  2  1  1  1  1  2 
#> [5737] 2  2  2  2  2  2  2  3  3  3  3  3  3  2  2  3  2  2  2  2  2  2  2  2 
#> [5761] 2  2  2  2  2  2  2  2  2  2  2  3  2  2  2  2  2  2  2  2  2  2  2  2 
#> [5785] 2  2  3  3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2 
#> [5809] 3  3  4  4  4  3  4  4  4  4  3  3  3  3  3  3  3  2  2  2  2  2  3  3 
#> [5833] 3  4  4  5  5  5  5  5  4  4  4  4  3  3  3  2  2  2  2  2  2  2  2  2 
#> [5857] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  2  2  2  2  2  2  1 
#> [5881] 1  1  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5905] 1  1  1  1  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  2  2  1 
#> [5929] 2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  2  2  1  1  1  2 
#> [5953] 2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2 
#> [5977] 2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  2  1  2  2 
#> [6001] 2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  1  1  1 
#> [6025] 1  1  1  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6049] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2 
#> [6073] 2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6097] 1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  3  2  2  2  2  2  1  1  1 
#> [6121] 1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1 
#> [6145] 1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1 
#> [6169] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6193] 1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1 
#> [6217] 1  1  1  1  1  1  1  1  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6241] 1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  1  1  1  1  1  1 
#> [6265] 1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  1  1  1  1  1  1 
#> [6289] 1  1  1  1  1  1  1  1  2  2  3  3  3  2  2  2  2  2  2  2  1  1  1  1 
#> [6313] 1  1  1  1  1  1  2  2  2  2  3  3  3  3  3  3  2  2  2  2  2  1  1  1 
#> [6337] 1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1 
#> [6361] 1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1 
#> [6385] 1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  1  1  1  1  1 
#> [6409] 1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1 
#> [6433] 1  1  1  1  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1 
#> [6457] 1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1 
#> [6481] 1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1 
#> [6505] 1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1 
#> [6529] 1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1 
#> [6553] 1  1  1  1  1  1  2  1  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1 
#> [6577] 1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1 
#> [6601] 1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1 
#> [6625] 1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1 
#> [6649] 1  1  1  1  1  1  2  2  3  3  3  3  3  3  2  2  2  2  2  1  1  1  1  1 
#> [6673] 2  2  2  2  2  2  3  3  3  3  3  2  2  2  2  2  2  2  2  1  1  1  1  1 
#> [6697] 1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1 
#> [6721] 1  1  1  1  1  2  2  2  2  2  2  2  3  3  2  2  2  2  2  1  1  1  1  1 
#> [6745] 1  1  2  2  2  2  2  2  3  3  3  3  3  2  2  2  2  2  1  1  1  1  1  1 
#> [6769] 1  2  2  2  2  3  3  4  4  4  3  3  3  3  2  2  2  2  2  1  1  1  1  2 
#> [6793] 2  2  2  3  3  3  3  4  4  5  4  4  4  3  3  2  2  2  2  2  2  1  1  2 
#> [6817] 2  2  2  2  2  3  4  4  4  4  4  4  3  3  3  3  2  2  2  2  2  2  2  2 
#> [6841] 2  2  3  3  4  5  5  5  4  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2 
#> [6865] 2  2  3  3  3  3  3  3  3  3  2  2  2  2  2  2  1  1  1  1  1  1  1  1 
#> [6889] 2  2  3  4  4  4  4  3  3  3  2  2  2  2  2  1  1  1  1  1  1  2  2  2 
#> [6913] 2  2  2  3  4  4  4  3  3  3  3  2  2  2  2  2  1  1  1  1  2  2  2  2 
#> [6937] 3  3  3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [6961] 2  2  2  2  2  3  3  3  2  2  2  2  2  1  1  1  1  1  1  2  2  2  2  2 
#> [6985] 2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  2  2  2  2 
#> [7009] 2  2  2  2  3  3  3  3  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7033] 1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  2  2  2  2 
#> [7057] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [7081] 2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2 
#> [7105] 2  2  2  2  2  2  2  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2 
#> [7129] 2  2  2  2  2  2  2  2  2  2  2  2  3  2  2  3  3  2  2  2  1  1  1  1 
#> [7153] 1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [7177] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  2  2  2  2  2 
#> [7201] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [7225] 2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  2  2  2  2  2 
#> [7249] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  2  2  2 
#> [7273] 2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7297] 2  1  2  2  3  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  2  2  2  2 
#> [7321] 2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  2  2  2 
#> [7345] 2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2 
#> [7369] 2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2 
#> [7393] 2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  3 
#> [7417] 3  3  2  2  3  3  3  2  2  4  4  4  5  5  4  4  4  3  3  3  3  3  3  3 
#> [7441] 3  2  2  2  2  2  2  2  2  2  1  2  1  2  2  2  2  2  2  2  2  2  2  2 
#> [7465] 2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  2  2  2 
#> [7489] 2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2 
#> [7513] 2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  3 
#> [7537] 2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  3 
#> [7561] 2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  2  2  2  2  2  3  3  2 
#> [7585] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  2  2  2  2  2  2  2 
#> [7609] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [7633] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [7657] 2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2 
#> [7681] 2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2 
#> [7705] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  2  2  2  2  2 
#> [7729] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [7753] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [7777] 2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  1  1 
#> [7801] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2 
#> [7825] 2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7849] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2 
#> [7873] 2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2 
#> [7897] 2  2  2  2  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  3  3  3  3  3 
#> [7921] 2  2  2  2  2  2  1  1  1  1  1  2  2  2  2  2  2  3  3  3  3  3  3  3 
#> [7945] 2  3  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  2  3  3  3  3  3  3 
#> [7969] 3  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  3  3  3  3 
#> [7993] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  2 
#> [8017] 2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2 
#> [8041] 2  2  2  2  2  2  2  2  2  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [8065] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  2 
#> [8089] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  2  2 
#> [8113] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  2  2  2  2 
#> [8137] 2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  2  2  2  2  2  2  3  2 
#> [8161] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [8185] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3 
#> [8209] 2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  2  2  2  2  2  2  2  2 
#> [8233] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [8257] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [8281] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  2  2 
#> [8305] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  2  2  2 
#> [8329] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [8353] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [8377] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1 
#> [8401] 1  1  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2 
#> [8425] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2 
#> [8449] 2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8473] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2 
#> [8497] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [8521] 2  2  2  2  2  1  2  2  2  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2 
#> [8545] 1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  3  2 
#> [8569] 2  2  2 
#> Levels: 1 2 3 4 5 6 7 8 9 10 +

# Calculate just the AQHI+
obs$date_utc |> AQHI(pm25_1hr_ugm3 = obs$pm25_1hr)
#> Warning: No non-missing NO2 / O3 data provided. Returning AQHI+ (PM2.5 only) instead of AQHI.
#> # A tibble: 8,759 × 18
#>    date                pm25_1hr_ugm3 o3_1hr_ppb no2_1hr_ppb pm25_3hr_ugm3
#>    <dttm>                      <dbl>      <dbl>       <dbl>         <dbl>
#>  1 2018-01-01 09:00:00           9.2         NA          NA            NA
#>  2 2018-01-01 10:00:00          12.1         NA          NA            NA
#>  3 2018-01-01 11:00:00          10.7         NA          NA            NA
#>  4 2018-01-01 12:00:00           9.3         NA          NA            NA
#>  5 2018-01-01 13:00:00           8           NA          NA            NA
#>  6 2018-01-01 14:00:00           6.6         NA          NA            NA
#>  7 2018-01-01 15:00:00           6           NA          NA            NA
#>  8 2018-01-01 16:00:00           6.1         NA          NA            NA
#>  9 2018-01-01 17:00:00           7.1         NA          NA            NA
#> 10 2018-01-01 18:00:00           7.9         NA          NA            NA
#> # ℹ 8,749 more rows
#> # ℹ 13 more variables: o3_3hr_ppb <dbl>, no2_3hr_ppb <dbl>, level <fct>,
#> #   AQHI <fct>, AQHI_plus <fct>, AQHI_plus_exceeds_AQHI <lgl>,
#> #   AQHI_pm25_ratio <dbl>, AQHI_o3_ratio <dbl>, AQHI_no2_ratio <dbl>,
#> #   colour <chr>, risk <fct>, high_risk_pop_message <chr>,
#> #   general_pop_message <chr>
obs$date_utc |> AQHI(pm25_1hr_ugm3 = obs$pm25_1hr, quiet = TRUE) # silence warning
#> # A tibble: 8,759 × 18
#>    date                pm25_1hr_ugm3 o3_1hr_ppb no2_1hr_ppb pm25_3hr_ugm3
#>    <dttm>                      <dbl>      <dbl>       <dbl>         <dbl>
#>  1 2018-01-01 09:00:00           9.2         NA          NA            NA
#>  2 2018-01-01 10:00:00          12.1         NA          NA            NA
#>  3 2018-01-01 11:00:00          10.7         NA          NA            NA
#>  4 2018-01-01 12:00:00           9.3         NA          NA            NA
#>  5 2018-01-01 13:00:00           8           NA          NA            NA
#>  6 2018-01-01 14:00:00           6.6         NA          NA            NA
#>  7 2018-01-01 15:00:00           6           NA          NA            NA
#>  8 2018-01-01 16:00:00           6.1         NA          NA            NA
#>  9 2018-01-01 17:00:00           7.1         NA          NA            NA
#> 10 2018-01-01 18:00:00           7.9         NA          NA            NA
#> # ℹ 8,749 more rows
#> # ℹ 13 more variables: o3_3hr_ppb <dbl>, no2_3hr_ppb <dbl>, level <fct>,
#> #   AQHI <fct>, AQHI_plus <fct>, AQHI_plus_exceeds_AQHI <lgl>,
#> #   AQHI_pm25_ratio <dbl>, AQHI_o3_ratio <dbl>, AQHI_no2_ratio <dbl>,
#> #   colour <chr>, risk <fct>, high_risk_pop_message <chr>,
#> #   general_pop_message <chr>

# Return French version
obs$date_utc |>
  AQHI(
    pm25_1hr_ugm3 = obs$pm25_1hr,
    o3_1hr_ppb = obs$o3_1hr,
    no2_1hr_ppb = obs$no2_1hr,
    language = "fr"
  )
#> # A tibble: 8,571 × 18
#>    date                pm25_1hr_ugm3 o3_1hr_ppb no2_1hr_ppb pm25_3hr_ugm3
#>    <dttm>                      <dbl>      <dbl>       <dbl>         <dbl>
#>  1 2018-01-01 09:00:00           9.2        0.7        18.1          NA  
#>  2 2018-01-01 10:00:00          12.1        0.3        17.5          NA  
#>  3 2018-01-01 11:00:00          10.7        0.2        17.8          10.7
#>  4 2018-01-01 12:00:00           9.3        0.2        18.1          10.7
#>  5 2018-01-01 13:00:00           8          1.1        16.6           9.3
#>  6 2018-01-01 14:00:00           6.6        1.4        15.6           8  
#>  7 2018-01-01 15:00:00           6          1.9        13.8           6.9
#>  8 2018-01-01 16:00:00           6.1        0.3        16.9           6.2
#>  9 2018-01-01 17:00:00           7.1        1          18.1           6.4
#> 10 2018-01-01 18:00:00           7.9        2.5        16.9           7  
#> # ℹ 8,561 more rows
#> # ℹ 13 more variables: o3_3hr_ppb <dbl>, no2_3hr_ppb <dbl>, level <fct>,
#> #   AQHI <fct>, AQHI_plus <fct>, AQHI_plus_exceeds_AQHI <lgl>,
#> #   AQHI_pm25_ratio <dbl>, AQHI_o3_ratio <dbl>, AQHI_no2_ratio <dbl>,
#> #   colour <chr>, risk <fct>, high_risk_pop_message <chr>,
#> #   general_pop_message <chr>
obs$date_utc |> AQHI(pm25_1hr_ugm3 = obs$pm25_1hr, language = "fr")
#> Warning: No non-missing NO2 / O3 data provided. Returning AQHI+ (PM2.5 only) instead of AQHI.
#> # A tibble: 8,759 × 18
#>    date                pm25_1hr_ugm3 o3_1hr_ppb no2_1hr_ppb pm25_3hr_ugm3
#>    <dttm>                      <dbl>      <dbl>       <dbl>         <dbl>
#>  1 2018-01-01 09:00:00           9.2         NA          NA            NA
#>  2 2018-01-01 10:00:00          12.1         NA          NA            NA
#>  3 2018-01-01 11:00:00          10.7         NA          NA            NA
#>  4 2018-01-01 12:00:00           9.3         NA          NA            NA
#>  5 2018-01-01 13:00:00           8           NA          NA            NA
#>  6 2018-01-01 14:00:00           6.6         NA          NA            NA
#>  7 2018-01-01 15:00:00           6           NA          NA            NA
#>  8 2018-01-01 16:00:00           6.1         NA          NA            NA
#>  9 2018-01-01 17:00:00           7.1         NA          NA            NA
#> 10 2018-01-01 18:00:00           7.9         NA          NA            NA
#> # ℹ 8,749 more rows
#> # ℹ 13 more variables: o3_3hr_ppb <dbl>, no2_3hr_ppb <dbl>, level <fct>,
#> #   AQHI <fct>, AQHI_plus <fct>, AQHI_plus_exceeds_AQHI <lgl>,
#> #   AQHI_pm25_ratio <dbl>, AQHI_o3_ratio <dbl>, AQHI_no2_ratio <dbl>,
#> #   colour <chr>, risk <fct>, high_risk_pop_message <chr>,
#> #   general_pop_message <chr>

# Don't allow AQHI+ override (not recommended)
obs$date_utc |>
  AQHI(
    pm25_1hr_ugm3 = obs$pm25_1hr,
    o3_1hr_ppb = obs$o3_1hr,
    no2_1hr_ppb = obs$no2_1hr,
    allow_aqhi_plus_override = FALSE
  )
#> # A tibble: 8,571 × 18
#>    date                pm25_1hr_ugm3 o3_1hr_ppb no2_1hr_ppb pm25_3hr_ugm3
#>    <dttm>                      <dbl>      <dbl>       <dbl>         <dbl>
#>  1 2018-01-01 09:00:00           9.2        0.7        18.1          NA  
#>  2 2018-01-01 10:00:00          12.1        0.3        17.5          NA  
#>  3 2018-01-01 11:00:00          10.7        0.2        17.8          10.7
#>  4 2018-01-01 12:00:00           9.3        0.2        18.1          10.7
#>  5 2018-01-01 13:00:00           8          1.1        16.6           9.3
#>  6 2018-01-01 14:00:00           6.6        1.4        15.6           8  
#>  7 2018-01-01 15:00:00           6          1.9        13.8           6.9
#>  8 2018-01-01 16:00:00           6.1        0.3        16.9           6.2
#>  9 2018-01-01 17:00:00           7.1        1          18.1           6.4
#> 10 2018-01-01 18:00:00           7.9        2.5        16.9           7  
#> # ℹ 8,561 more rows
#> # ℹ 13 more variables: o3_3hr_ppb <dbl>, no2_3hr_ppb <dbl>, level <fct>,
#> #   AQHI <fct>, AQHI_plus <fct>, AQHI_plus_exceeds_AQHI <lgl>,
#> #   AQHI_pm25_ratio <dbl>, AQHI_o3_ratio <dbl>, AQHI_no2_ratio <dbl>,
#> #   colour <chr>, risk <fct>, high_risk_pop_message <chr>,
#> #   general_pop_message <chr>