Calculate the Canadian AQHI+ from hourly PM2.5 concentrations

The Canadian AQHI+ is a modification of the Canadian Air Quality Health Index (AQHI). AQHI+ only uses fine particulate matter (PM2.5) instead of the combination of PM2.5, ozone (O3), and nitrogen dioxide (NO2). Unlike the AQHI which uses 3-hourly mean averages, AQHI+ is calculated using hourly mean averages. The AQHI+ overrides the AQHI if it exceeds the AQHI for a particular hour.

AQHI+ categorizes hourly PM2.5 concentrations into 10 equal bins from 0 to 100 μg m-3 (AQHI+ 1-10), and assigns "+" to values greater than 100 μg m-3. The risk categories match the Canadian AQHI and share the same health messaging:

Low: AQHI+ of 1-3, or a concentration of 0-30 μg m-3

Moderate: AQHI+ of 4-6, or a concentration of 30.1-60 μg m-3

High: AQHI+ of 7-10, or a concentration of 60.1-100 μg m-3

Very High: AQHI+ of +, or a concentration above 100 μg m-3

The AQHI+ was originally published by Yao et al (2019): doi:10.17269/s41997-019-00237-w , and has been adopted by all* Canadian provinces/territories as of 2024. (*except Québec where they use the AQI instead of the AQHI/AQHI+).

Quote from Yao et al (2019):

This algorithm is a linear extrapolation based on two anchor points:

the AQHI-Plus value should reach 7 (high risk category), when 1-h PM2.5 concentrations are over 60 ug/m3. This is consistent with [...] the analyses of the likelihood of exceeding the 25 ug/m3 24-h threshold value; and

the AQHI-Plus value should reach 4 (moderate risk category) when the 1-h PM2.5 concentrations are over 30 ug/m3, which is the threshold for odour perception of burning softwood (Kistler et al. 2012: doi:10.1016/j.atmosenv.2012.01.044 ), the major type of forest species in [British Columbia].

[...]

These conclusions suggest that the AQHI is most indicative of the mortality and circulatory risks during periods affected by wildfire smoke. However, the [...] AQHI-Plus is most indicative of the respiratory risks during high-intensity fire periods, particularly for people with asthma. This is consistent with the rationale behind the AQHI formulation and with existing health evidence related to the effects of wildfire smoke.

Usage

AQHI_plus(
  pm25_1hr_ugm3,
  min_allowed_pm25 = 0,
  detailed = TRUE,
  language = "en"
)

Arguments

pm25_1hr_ugm3

A numeric vector of hourly mean PM2.5 concentrations (units = μg m-3). Must be a single value or the same length as dates or a single value (which will be repeated).

min_allowed_pm25

(Optional). A single numeric value indicating the minimum allowed concentration. All values in pm25_1hr_ugm3 less than this will be replaced with NA. Default is 0 μg m-3.

detailed

(Optional). A single logical value indicating if a tibble with levels, risk categories, health messages, etc should be returned. If FALSE only the levels will be returned. Default is TRUE.

language

(Optional). A single character value indicating the language to use for risk levels and health messaging. Must be either "en" (English) or "fr" (French). Not case sensitive. Default is "en".

Value

If detailed is TRUE:

A tibble (data.frame) with columns detailing the AQHI+, risk level, health messages, etc

If detailed is FALSE:

A factor vector of AQHI+ levels with length(pm25_1hr_ugm3) elements

References

Yao et al (2019): doi:10.17269/s41997-019-00237-w

Environment and Climate Change Canada: https://www.canada.ca/en/environment-climate-change/services/air-quality-health-index/about.html

Examples

# Hourly pm2.5 concentrations
pm25 <- example_obs$pm25_1hr[example_obs$site_id == 1]
# Calculate the detailed AQHI+ (tibble with risk levels etc)
AQHI_plus(pm25)
#> # A tibble: 8,571 × 6
#>    pm25_1hr_ugm3 level colour  risk  high_risk_pop_message   general_pop_message
#>            <dbl> <fct> <chr>   <fct> <chr>                   <chr>              
#>  1           9.2 1     #21C6F5 Low   Enjoy your usual activ… Ideal air quality …
#>  2          12.1 2     #189ACA Low   Enjoy your usual activ… Ideal air quality …
#>  3          10.7 2     #189ACA Low   Enjoy your usual activ… Ideal air quality …
#>  4           9.3 1     #21C6F5 Low   Enjoy your usual activ… Ideal air quality …
#>  5           8   1     #21C6F5 Low   Enjoy your usual activ… Ideal air quality …
#>  6           6.6 1     #21C6F5 Low   Enjoy your usual activ… Ideal air quality …
#>  7           6   1     #21C6F5 Low   Enjoy your usual activ… Ideal air quality …
#>  8           6.1 1     #21C6F5 Low   Enjoy your usual activ… Ideal air quality …
#>  9           7.1 1     #21C6F5 Low   Enjoy your usual activ… Ideal air quality …
#> 10           7.9 1     #21C6F5 Low   Enjoy your usual activ… Ideal air quality …
#> # ℹ 8,561 more rows
# Or just the AQHI+ (returned as a factor)
AQHI_plus(pm25, detailed = FALSE)
#>    [1] 1  2  2  1  1  1  1  1  1  1  1  1  2  2  2  2  3  2  1  1  1  1  1  1 
#>   [25] 1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  1  2  2  2  2  2  1  1  1 
#>   [49] 1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  1  1  1  1  1  1  1  1 
#>   [73] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>   [97] 1  1  1  1  1  1  1  1  1  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [121] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [145] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [169] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [193] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [217] 1  1  1  1  1  1  1  1  1  1  2  2  2  1  1  1  1  1  1  1  1  1  1  1 
#>  [241] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [265] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  1  1  1 
#>  [289] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [313] 1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  1  1  1  1  1  1  1  1  1 
#>  [337] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [361] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [385] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [409] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [433] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [457] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [481] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [505] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [529] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [553] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [577] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [601] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  1  1  1  1  1  1 
#>  [625] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [649] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [673] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [697] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [721] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [745] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [769] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [793] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [817] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [841] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [865] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [889] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [913] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [937] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [961] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#>  [985] 1  1  1  1  1  1  1  2  1  1  1  1  1  2  1  1  1  1  1  1  1  1  1  1 
#> [1009] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1033] 1  2  1  1  1  1  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1 
#> [1057] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1081] 1  1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  1  1  1  1  1  1  1  1 
#> [1105] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1129] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1153] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  1  1  1  1  1 
#> [1177] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1201] 1  1  1  1  1  1  1  1  1  1  1  3  2  1  1  1  1  1  1  2  1  1  1  1 
#> [1225] 1  1  1  1  1  1  2  2  2  1  1  2  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1249] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1273] 1  1  1  1  1  1  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1297] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2 
#> [1321] 2  2  1  2  2  2  1  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1345] 1  1  1  1  1  1  1  1  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1369] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1393] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1417] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1441] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1465] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  4  3  2 
#> [1489] 2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1513] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1537] 1  1  1  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1561] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1585] 1  1  1  1  1  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  1 
#> [1609] 1  1  1  1  1  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1633] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1657] 1  1  1  1  1  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  1 
#> [1681] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1705] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1729] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1753] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1777] 1  1  1  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1801] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  1 
#> [1825] 2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1849] 1  1  1  1  1  1  2  2  2  2  1  1  1  1  1  1  1  1  2  1  1  1  1  1 
#> [1873] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1897] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1921] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1945] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1969] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [1993] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2017] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2041] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2065] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2089] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2113] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2137] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2161] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2185] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  1  1  1  1  1 
#> [2209] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2233] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2257] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2281] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2305] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2329] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2353] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2377] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2401] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2425] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2449] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2473] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2497] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2521] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2545] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2569] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2593] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2617] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2641] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2665] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2689] 1  1  2  2  2  2  1  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1 
#> [2713] 1  1  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2737] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2761] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2785] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2809] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  1  1  1  1  2 
#> [2833] 2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2857] 1  1  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2 
#> [2881] 2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2905] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2929] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2953] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [2977] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3001] 2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3025] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3049] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3073] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3097] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2 
#> [3121] 2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2 
#> [3145] 2  2  2  2  2  1  1  1  1  1  1  1  1  2  2  2  1  1  1  1  1  2  2  2 
#> [3169] 2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3193] 1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  1  1  1  1  1  1  1  1  1 
#> [3217] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3241] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  1  1 
#> [3265] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3289] 1  1  1  1  1  1  1  1  4  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3313] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3337] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3361] 1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3385] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  1  1 
#> [3409] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3433] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3457] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3481] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3505] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3529] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3553] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3577] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3601] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3625] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3649] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3673] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3697] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3721] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3745] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3769] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3793] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3817] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3841] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3865] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3889] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [3913] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  1  1  1 
#> [3937] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  1  1 
#> [3961] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  2  1  1 
#> [3985] 1  1  1  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4009] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  1  1 
#> [4033] 2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4057] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4081] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4105] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4129] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4153] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4177] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4201] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4225] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4249] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4273] 1  1  1  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4297] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4321] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4345] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4369] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4393] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4417] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4441] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4465] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4489] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4513] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4537] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4561] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4585] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4609] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  1 
#> [4633] 1  1  1  1  1  1  1  1  1  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4657] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4681] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4705] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4729] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  6  1  1  1 
#> [4753] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  4  1  1  1  1 
#> [4777] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4801] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4825] 1  1  1  1  1  1  1  1  2  2  2  1  2  1  1  1  1  1  1  1  1  1  1  1 
#> [4849] 2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  2  1  2  2 
#> [4873] 2  2  2  2  2  2  2  2  2  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [4897] 2  2  2  1  1  1  1  1  1  1  2  2  2  2  1  1  1  1  1  2  1  1  1  2 
#> [4921] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [4945] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1 
#> [4969] 1  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [4993] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5017] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5041] 1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  1  1  1  1  1  1  1  1  1 
#> [5065] 1  1  1  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1 
#> [5089] 1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  1  1  1  2  2  1  1  1  1 
#> [5113] 1  1  1  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5137] 1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  2  2  1  2  2  2  2 
#> [5161] 2  1  1  1  1  2  2  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [5185] 2  3  2  3  5  4  3  2  1  2  3  2  2  2  1  1  1  1  1  1  1  1  1  1 
#> [5209] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5233] 1  1  1  1  1  1  1  1  1  2  2  2  2  1  1  1  1  2  2  2  2  2  3  3 
#> [5257] 3  4  4  5  5  5  5  5  5  6  4  4  4  4  4  8  9  8  7  7  7  7  7  7 
#> [5281] 7  8  8  8  8  8  7  6  7  7  8  8  8  7  7  7  7  7  6  6  6  5  5  6 
#> [5305] 6  7  6  5  6  6  5  4  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3  3 
#> [5329] 3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3 
#> [5353] 2  2  2  2  2  3  2  3  2  2  2  2  1  2  2  2  1  1  1  1  2  2  2  2 
#> [5377] 2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [5401] 2  2  3  5  3  4  3  4  6  6  6  6  5  5  6  6  5  4  5  7  8  8  9  + 
#> [5425] +  +  +  +  +  10 10 9  9  9  8  8  8  7  6  6  6  5  4  4  4  3  3  3 
#> [5449] 4  3  4  6  9  10 8  6  6  7  8  9  8  6  6  6  6  6  7  6  6  5  6  6 
#> [5473] 6  6  6  9  10 10 +  +  +  +  +  +  +  +  +  +  +  +  +  9  5  4  3  4 
#> [5497] 5  5  6  5  4  4  3  3  3  3  2  2  2  2  2  2  2  3  3  3  3  4  4  3 
#> [5521] 3  3  3  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  3  3 
#> [5545] 3  3  3  4  4  4  4  3  4  4  4  4  4  5  5  4  4  5  4  4  4  4  4  3 
#> [5569] 3  2  2  1  2  2  2  2  2  2  2  2  2  1  1  1  2  1  1  1  1  1  1  1 
#> [5593] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5617] 1  1  1  1  2  2  2  2  2  2  2  1  1  1  2  2  2  2  2  2  2  2  2  2 
#> [5641] 3  3  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5665] 1  1  1  1  1  1  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5689] 1  1  1  1  1  1  1  1  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5713] 1  1  1  1  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5737] 1  1  1  2  1  1  1  2  2  1  1  1  1  1  2  3  1  1  1  1  1  1  1  1 
#> [5761] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5785] 1  1  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  2  2  1 
#> [5809] 2  2  2  2  2  2  4  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [5833] 2  3  4  4  3  3  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5857] 1  1  1  1  1  2  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  1  1 
#> [5881] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5905] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5929] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5953] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [5977] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6001] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  1  1  1 
#> [6025] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6049] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6073] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6097] 1  1  1  1  1  1  1  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6121] 1  1  1  1  1  1  1  1  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6145] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6169] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6193] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6217] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6241] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6265] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6289] 1  1  1  1  1  1  1  1  2  2  3  2  2  1  1  1  1  1  1  1  1  1  1  1 
#> [6313] 1  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1 
#> [6337] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6361] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6385] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  1  1  1  1  1  1  1  1 
#> [6409] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6433] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6457] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6481] 1  1  1  1  1  1  1  2  1  1  1  1  1  2  2  1  1  1  1  1  1  1  1  1 
#> [6505] 1  1  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6529] 1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  1  2  2  1  1  1  1  1 
#> [6553] 1  1  1  1  1  1  2  1  2  1  1  2  2  1  1  1  1  1  1  1  1  1  1  1 
#> [6577] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6601] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6625] 1  1  1  1  1  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6649] 1  1  1  1  1  1  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1 
#> [6673] 2  2  2  2  2  2  3  3  3  2  1  1  1  2  2  2  2  2  2  1  1  1  1  1 
#> [6697] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6721] 1  1  1  1  1  1  1  1  1  1  1  2  1  2  1  1  1  1  1  1  1  1  1  1 
#> [6745] 1  1  1  1  2  1  1  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1 
#> [6769] 1  1  1  1  1  1  1  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1 
#> [6793] 2  2  2  2  1  1  2  2  3  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1 
#> [6817] 1  1  1  1  1  2  2  3  3  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1 
#> [6841] 1  2  2  2  3  3  4  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1 
#> [6865] 1  2  3  3  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1 
#> [6889] 1  1  2  2  2  2  2  2  2  1  2  1  1  1  1  1  1  1  1  1  1  2  2  1 
#> [6913] 1  1  2  2  3  2  2  2  2  2  2  1  1  1  1  1  1  1  1  1  2  2  2  2 
#> [6937] 2  2  2  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2 
#> [6961] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [6985] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7009] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7033] 1  1  1  1  1  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7057] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7081] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2 
#> [7105] 2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7129] 1  1  1  1  1  1  1  1  1  1  1  1  3  2  2  3  3  2  2  1  1  1  1  1 
#> [7153] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7177] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7201] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7225] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7249] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7273] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7297] 1  1  1  2  2  2  2  2  2  2  1  2  2  2  2  1  1  1  1  1  1  1  2  2 
#> [7321] 2  2  1  1  2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7345] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2 
#> [7369] 2  2  2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7393] 2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2 
#> [7417] 2  2  2  2  3  3  3  2  2  4  4  4  5  5  4  4  3  2  1  1  2  2  2  2 
#> [7441] 1  1  1  2  2  2  1  1  1  1  1  2  1  1  2  1  2  1  1  1  1  1  1  1 
#> [7465] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7489] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2 
#> [7513] 2  2  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  3 
#> [7537] 1  1  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2 
#> [7561] 2  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2 
#> [7585] 2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1  2  1  1  1 
#> [7609] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7633] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7657] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7681] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7705] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7729] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7753] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7777] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7801] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7825] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7849] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2 
#> [7873] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7897] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [7921] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  1  1  2  1 
#> [7945] 1  3  1  1  1  1  1  1  1  1  1  1  2  2  2  2  1  1  2  2  2  2  2  2 
#> [7969] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  1 
#> [7993] 1  2  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  3  1 
#> [8017] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8041] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8065] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8089] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8113] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8137] 1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8161] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8185] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8209] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8233] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8257] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8281] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  1  1  1  1 
#> [8305] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8329] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8353] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8377] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8401] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8425] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8449] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8473] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8497] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8521] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> [8545] 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  1  3  2 
#> [8569] 1  2  2 
#> Levels: 1 2 3 4 5 6 7 8 9 10 +