Get AQHI colours for each AQHI level

Every AQHI level (1-10, "+", and NA) is assigned a hexidecimal colour code for visualization.

These colours are defined by Environment and Climate Change Canada such that:

Low AQHI (1-3) are light to dark blue

Moderate AQHI (4-6) are yellow to orange

High AQHI (7-10) are light to dark red

Very High AQHI (+) is darker red

Missing AQHI (NA) is light grey

Usage

get_aqhi_colours(values = c(1:10, "+", NA), types = "aqhi")

Arguments

values

(Optional). A vector of AQHI levels (1-10, "+", or NA) if type is "aqhi" OR A vector of hourly PM2.5 concentrations (ug/m3) if type is "pm25_1hr". Default is all AQHI levels.

types

(Optional). A single character value, or a vector the same length as values, indicating the type(s) of values in values. Must all be within either "aqhi" or "pm25_1hr". Default is "aqhi".

Value

A character vector of hexidecimal codes correspoding to each level of aqhi_levels.

References

Environment and Climate Change Canada: https://www.canada.ca/en/environment-climate-change/services/air-quality-health-index/about.html

Examples

# Get AQHI colours for all AQHI levels
get_aqhi_colours()
#> Warning: NAs introduced by coercion
#>  [1] "#21C6F5" "#189ACA" "#0D6797" "#FFFD37" "#FFCC2E" "#FE9A3F" "#FD6769"
#>  [8] "#FF3B3B" "#FF0101" "#CB0713" "#bbbbbb" "#bbbbbb"

# Get AQHI colours for observation data
hourly_pm25_ugm3 <- sample(1:100, 50, replace = TRUE)
aqhi_levels <- hourly_pm25_ugm3 |>
  AQHI_plus(detailed = FALSE)
aqhi_levels |> get_aqhi_colours()
#>  [1] "#FE9A3F" "#FFFD37" "#FE9A3F" "#21C6F5" "#FE9A3F" "#CB0713" "#189ACA"
#>  [8] "#21C6F5" "#189ACA" "#FF3B3B" "#189ACA" "#FD6769" "#FF3B3B" "#FF0101"
#> [15] "#0D6797" "#FD6769" "#FFFD37" "#FF0101" "#21C6F5" "#FF3B3B" "#FE9A3F"
#> [22] "#FFCC2E" "#FFFD37" "#189ACA" "#FF0101" "#189ACA" "#FFFD37" "#FF0101"
#> [29] "#CB0713" "#FD6769" "#FE9A3F" "#0D6797" "#21C6F5" "#FFCC2E" "#CB0713"
#> [36] "#FE9A3F" "#FD6769" "#FE9A3F" "#FD6769" "#FD6769" "#FFFD37" "#FFCC2E"
#> [43] "#FFCC2E" "#FFFD37" "#189ACA" "#FF0101" "#CB0713" "#FF3B3B" "#CB0713"
#> [50] "#21C6F5"

# The same but with PM2.5 provided
hourly_pm25_ugm3 |> get_aqhi_colours(types = "pm25_1hr")
#>  [1] "#FE9A3F" "#FFFD37" "#FE9A3F" "#21C6F5" "#FE9A3F" "#CB0713" "#189ACA"
#>  [8] "#21C6F5" "#189ACA" "#FF3B3B" "#189ACA" "#FD6769" "#FF3B3B" "#FF0101"
#> [15] "#0D6797" "#FD6769" "#FFFD37" "#FF0101" "#21C6F5" "#FF3B3B" "#FE9A3F"
#> [22] "#FFCC2E" "#FFFD37" "#189ACA" "#FF0101" "#189ACA" "#FFFD37" "#FF0101"
#> [29] "#CB0713" "#FD6769" "#FE9A3F" "#0D6797" "#21C6F5" "#FFCC2E" "#CB0713"
#> [36] "#FE9A3F" "#FD6769" "#FE9A3F" "#FD6769" "#FD6769" "#FFFD37" "#FFCC2E"
#> [43] "#FFCC2E" "#FFFD37" "#189ACA" "#FF0101" "#CB0713" "#FF3B3B" "#CB0713"
#> [50] "#21C6F5"

# Or even a mix
values <- c(aqhi_levels, hourly_pm25_ugm3)
types <-  rep("aqhi", length(aqhi_levels)) |>
  c(rep("pm25_1hr", length(hourly_pm25_ugm3)))

values |> get_aqhi_colours(types = types)
#>   [1] "#FE9A3F" "#FFFD37" "#FE9A3F" "#21C6F5" "#FE9A3F" "#CB0713" "#189ACA"
#>   [8] "#21C6F5" "#189ACA" "#FF3B3B" "#189ACA" "#FD6769" "#FF3B3B" "#FF0101"
#>  [15] "#0D6797" "#FD6769" "#FFFD37" "#FF0101" "#21C6F5" "#FF3B3B" "#FE9A3F"
#>  [22] "#FFCC2E" "#FFFD37" "#189ACA" "#FF0101" "#189ACA" "#FFFD37" "#FF0101"
#>  [29] "#CB0713" "#FD6769" "#FE9A3F" "#0D6797" "#21C6F5" "#FFCC2E" "#CB0713"
#>  [36] "#FE9A3F" "#FD6769" "#FE9A3F" "#FD6769" "#FD6769" "#FFFD37" "#FFCC2E"
#>  [43] "#FFCC2E" "#FFFD37" "#189ACA" "#FF0101" "#CB0713" "#FF3B3B" "#CB0713"
#>  [50] "#21C6F5" "#FE9A3F" "#FFFD37" "#FE9A3F" "#21C6F5" "#FE9A3F" "#CB0713"
#>  [57] "#189ACA" "#21C6F5" "#189ACA" "#FF3B3B" "#189ACA" "#FD6769" "#FF3B3B"
#>  [64] "#FF0101" "#0D6797" "#FD6769" "#FFFD37" "#FF0101" "#21C6F5" "#FF3B3B"
#>  [71] "#FE9A3F" "#FFCC2E" "#FFFD37" "#189ACA" "#FF0101" "#189ACA" "#FFFD37"
#>  [78] "#FF0101" "#CB0713" "#FD6769" "#FE9A3F" "#0D6797" "#21C6F5" "#FFCC2E"
#>  [85] "#CB0713" "#FE9A3F" "#FD6769" "#FE9A3F" "#FD6769" "#FD6769" "#FFFD37"
#>  [92] "#FFCC2E" "#FFCC2E" "#FFFD37" "#189ACA" "#FF0101" "#CB0713" "#FF3B3B"
#>  [99] "#CB0713" "#21C6F5"