Get health messages for AQHI+ risk categories

Health messages for at risk population and general population are available for each AQHI health risk category (see get_risk_category).

English messages:

Risk	AQHI	At Risk Population	General Population

Low	1 - 3	Enjoy your usual outdoor activities.	Ideal air quality for outdoor activities.

Moderate	4 - 6	Consider reducing or rescheduling strenuous activities outdoors if you are experiencing symptoms.	No need to modify your usual outdoor activities unless you experience symptoms such as coughing and throat irritation.

High	7 - 10	Reduce or reschedule strenuous activities outdoors. Children and the elderly should also take it easy.	Consider reducing or rescheduling strenuous activities outdoors if you experience symptoms such as coughing and throat irritation.

Very High	>10	Avoid strenuous activities outdoors. Children and the elderly should also avoid outdoor physical exertion.	Reduce or reschedule strenuous activities outdoors, especially if you experience symptoms such as coughing and throat irritation.

See Environment and Climate Change Canada's website for more information.

Usage

get_health_messages(
  risk_categories = c("Low", "Moderate", "High", "Very High"),
  language = "en"
)

Arguments

risk_categories

A factor or character vector of AQHI risk categories (Low, Moderate, High, Very High).

language

(Optional). A single character value indicating the language to use for risk levels and health messaging. Must be either "en" (English) or "fr" (French). Not case sensitive. Default is "en".

Value

A tibble of health messages (for at risk population and general population) for the provided risk categories.

Examples

# Get health messages for all risk categories
get_health_messages()
#> # A tibble: 4 × 3
#>   risk_category high_risk_pop_message                        general_pop_message
#>   <chr>         <chr>                                        <chr>              
#> 1 Low           Enjoy your usual activities.                 Ideal air quality …
#> 2 Moderate      Consider reducing or rescheduling activitie… No need to modify …
#> 3 High          Reduce or reschedule activities outdoors.    Consider reducing …
#> 4 Very High     Avoid strenuous activity outdoors.           Reduce or reschedu…

# The same, but en Francais
get_health_messages(language = "fr")
#> # A tibble: 4 × 3
#>   risk_category high_risk_pop_message general_pop_message
#>   <chr>         <chr>                 <chr>              
#> 1 NA            NA                    NA                 
#> 2 NA            NA                    NA                 
#> 3 NA            NA                    NA                 
#> 4 NA            NA                    NA                 

# Get health messages for some observations
hourly_pm25_ugm3 <- sample(1:100, 50, replace = TRUE)
risk_categories <- hourly_pm25_ugm3 |>
  AQHI_plus(detailed = FALSE) |>
  get_risk_category()
risk_categories |> get_health_messages()
#> # A tibble: 50 × 3
#>    risk_category high_risk_pop_message                       general_pop_message
#>    <chr>         <chr>                                       <chr>              
#>  1 High          Reduce or reschedule activities outdoors.   Consider reducing …
#>  2 High          Reduce or reschedule activities outdoors.   Consider reducing …
#>  3 High          Reduce or reschedule activities outdoors.   Consider reducing …
#>  4 Low           Enjoy your usual activities.                Ideal air quality …
#>  5 Moderate      Consider reducing or rescheduling activiti… No need to modify …
#>  6 High          Reduce or reschedule activities outdoors.   Consider reducing …
#>  7 High          Reduce or reschedule activities outdoors.   Consider reducing …
#>  8 Moderate      Consider reducing or rescheduling activiti… No need to modify …
#>  9 Low           Enjoy your usual activities.                Ideal air quality …
#> 10 High          Reduce or reschedule activities outdoors.   Consider reducing …
#> # ℹ 40 more rows