Calculate the Canadian AQHI from hourly PM2.5, O3, and NO2 concentrations

The Canadian Air Quality Health Index (AQHI) is a measure of the health risk associated with exposure to ambient air pollution. It combines the impacts of fine particulate matter (PM2.5), ozone (O3), and nitrogen dioxide (NO2) on a scale from 1-10 (+).

AQHI formula: 10⁄10.4 × 100 × (e0.000537 * O3 - 1) + (e0.000871 * NO2 - 1) + (e0.000487 * PM2.5 - 1)

The AQHI was originally published by Stieb et al. (2008): doi:10.3155/1047-3289.58.3.435 , and is used by all Canadian provinces and territories except for Québec, which uses the AQI instead. The AQHI is calculated for each hour using 3-hour rolling mean concentrations of PM2.5, O3, and NO2. However, the AQHI is overriden by the AQHI+ for a given hour if that AQHI+ level is higher. The AQHI+ is a modification of the AQHI that only uses PM2.5 and is calculated using hourly means (see AQHI_plus).

Usage

AQHI(
  dates,
  pm25_1hr_ugm3,
  o3_1hr_ppb = NA_real_,
  no2_1hr_ppb = NA_real_,
  allow_aqhi_plus_override = TRUE,
  detailed = TRUE,
  language = "en",
  quiet = FALSE
)

Arguments

dates

A POSIXct (date-time) vector corresponding to hourly pollutant observations. Date gaps will be filled with NA's automatically as needed when calculating 3-hour averages.

pm25_1hr_ugm3

A numeric vector of hourly mean PM2.5 concentrations (units = μg m-3). Must be a single value or the same length as dates or a single value (which will be repeated).

o3_1hr_ppb

(Optional). A numeric vector of hourly mean ozone (O3) concentrations (units = ppb). Must be a single value or the same length as dates or a single value (which will be repeated). If all are NA (the default) AQHI won't be calculated and instead AQHI+ (see AQHI_plus) will be used.

no2_1hr_ppb

(Optional). A numeric vector of hourly mean nitrogen dioxide (NO2) concentrations (units = ppb). Must be a single value or the same length as dates or a single value (which will be repeated). If all are NA (the default) AQHI won't be calculated and instead AQHI+ (see AQHI_plus) will used.

allow_aqhi_plus_override

(Optional). A single logical value indicating if the AQHI+ should be allowed to override the AQHI if it exceeds the AQHI for a particular hour. Default is TRUE.

detailed

(Optional). A single logical value indicating if a tibble with levels, risk categories, health messages, etc should be returned. If FALSE only the levels will be returned. Default is TRUE.

language

(Optional). A single character value indicating the language to use for risk levels and health messaging. Must be either "en" (English) or "fr" (French). Not case sensitive. Default is "en".

quiet

(Optional). A single logical (TRUE/FALSE) value indicating if non-critical warnings/messages should be hidden. Default is FALSE.

Value

If detailed = TRUE:

A tibble (data.frame) with a detailed summary of the AQHI, risk levels, health messages, etc and the same number of rows as length(dates)

If detailed = FALSE:

a factor vector of AQHI levels with length(dates) elements

References

Stieb et al. (2008): doi:10.3155/1047-3289.58.3.435

Environment and Climate Change Canada: https://www.canada.ca/en/environment-climate-change/services/air-quality-health-index/about.html

Examples

obs <- data.frame(
  date = seq(
    as.POSIXct("2024-01-01 00:00:00"),
    as.POSIXct("2024-01-01 23:00:00"), "1 hours"
  ),
  pm25 = sample(1:150, 24), o3 = sample(1:150, 24), no2 = sample(1:150, 24)
)

# Calculate the AQHI
obs$date |>
  AQHI(
    pm25_1hr_ugm3 = obs$pm25,
    o3_1hr_ppb = obs$o3,
    no2_1hr_ppb = obs$no2
  )
#> # A tibble: 24 × 18
#>    date                pm25_1hr_ugm3 o3_1hr_ppb no2_1hr_ppb pm25_3hr_ugm3
#>    <dttm>                      <dbl>      <dbl>       <dbl>         <dbl>
#>  1 2024-01-01 00:00:00            63         17          12          NA  
#>  2 2024-01-01 01:00:00           102         22          41          NA  
#>  3 2024-01-01 02:00:00           111         95          42          92  
#>  4 2024-01-01 03:00:00            47        129          66          86.7
#>  5 2024-01-01 04:00:00             5         70         102          54.3
#>  6 2024-01-01 05:00:00            24         65         103          25.3
#>  7 2024-01-01 06:00:00            79         25          26          36  
#>  8 2024-01-01 07:00:00            77          8          60          60  
#>  9 2024-01-01 08:00:00           130        108          14          95.3
#> 10 2024-01-01 09:00:00            55         83           5          87.3
#> # ℹ 14 more rows
#> # ℹ 13 more variables: o3_3hr_ppb <dbl>, no2_3hr_ppb <dbl>, level <fct>,
#> #   AQHI <fct>, AQHI_plus <fct>, AQHI_plus_exceeds_AQHI <lgl>,
#> #   AQHI_pm25_ratio <dbl>, AQHI_o3_ratio <dbl>, AQHI_no2_ratio <dbl>,
#> #   colour <chr>, risk <fct>, high_risk_pop_message <chr>,
#> #   general_pop_message <chr>

# Return just the AQHI levels
obs$date |>
  AQHI(
    pm25_1hr_ugm3 = obs$pm25,
    o3_1hr_ppb = obs$o3,
    no2_1hr_ppb = obs$no2,
    detailed = FALSE
  )
#>  [1] 7  +  +  +  +  +  +  10 +  10 +  +  +  10 10 +  +  +  +  +  10 8  +  + 
#> Levels: 1 2 3 4 5 6 7 8 9 10 +

# Calculate just the AQHI+
obs$date |> AQHI(pm25_1hr_ugm3 = obs$pm25)
#> Warning: No non-missing NO2 / O3 data provided. Returning AQHI+ (PM2.5 only) instead of AQHI.
#> # A tibble: 24 × 18
#>    date                pm25_1hr_ugm3 o3_1hr_ppb no2_1hr_ppb pm25_3hr_ugm3
#>    <dttm>                      <dbl>      <dbl>       <dbl>         <dbl>
#>  1 2024-01-01 00:00:00            63         NA          NA            NA
#>  2 2024-01-01 01:00:00           102         NA          NA            NA
#>  3 2024-01-01 02:00:00           111         NA          NA            NA
#>  4 2024-01-01 03:00:00            47         NA          NA            NA
#>  5 2024-01-01 04:00:00             5         NA          NA            NA
#>  6 2024-01-01 05:00:00            24         NA          NA            NA
#>  7 2024-01-01 06:00:00            79         NA          NA            NA
#>  8 2024-01-01 07:00:00            77         NA          NA            NA
#>  9 2024-01-01 08:00:00           130         NA          NA            NA
#> 10 2024-01-01 09:00:00            55         NA          NA            NA
#> # ℹ 14 more rows
#> # ℹ 13 more variables: o3_3hr_ppb <dbl>, no2_3hr_ppb <dbl>, level <fct>,
#> #   AQHI <fct>, AQHI_plus <fct>, AQHI_plus_exceeds_AQHI <lgl>,
#> #   AQHI_pm25_ratio <dbl>, AQHI_o3_ratio <dbl>, AQHI_no2_ratio <dbl>,
#> #   colour <chr>, risk <fct>, high_risk_pop_message <chr>,
#> #   general_pop_message <chr>
obs$date |> AQHI(pm25_1hr_ugm3 = obs$pm25, quiet = TRUE) # silence warning
#> # A tibble: 24 × 18
#>    date                pm25_1hr_ugm3 o3_1hr_ppb no2_1hr_ppb pm25_3hr_ugm3
#>    <dttm>                      <dbl>      <dbl>       <dbl>         <dbl>
#>  1 2024-01-01 00:00:00            63         NA          NA            NA
#>  2 2024-01-01 01:00:00           102         NA          NA            NA
#>  3 2024-01-01 02:00:00           111         NA          NA            NA
#>  4 2024-01-01 03:00:00            47         NA          NA            NA
#>  5 2024-01-01 04:00:00             5         NA          NA            NA
#>  6 2024-01-01 05:00:00            24         NA          NA            NA
#>  7 2024-01-01 06:00:00            79         NA          NA            NA
#>  8 2024-01-01 07:00:00            77         NA          NA            NA
#>  9 2024-01-01 08:00:00           130         NA          NA            NA
#> 10 2024-01-01 09:00:00            55         NA          NA            NA
#> # ℹ 14 more rows
#> # ℹ 13 more variables: o3_3hr_ppb <dbl>, no2_3hr_ppb <dbl>, level <fct>,
#> #   AQHI <fct>, AQHI_plus <fct>, AQHI_plus_exceeds_AQHI <lgl>,
#> #   AQHI_pm25_ratio <dbl>, AQHI_o3_ratio <dbl>, AQHI_no2_ratio <dbl>,
#> #   colour <chr>, risk <fct>, high_risk_pop_message <chr>,
#> #   general_pop_message <chr>

# Return French version
obs$date |>
  AQHI(
    pm25_1hr_ugm3 = obs$pm25,
    o3_1hr_ppb = obs$o3,
    no2_1hr_ppb = obs$no2,
    language = "fr"
  )
#> # A tibble: 24 × 18
#>    date                pm25_1hr_ugm3 o3_1hr_ppb no2_1hr_ppb pm25_3hr_ugm3
#>    <dttm>                      <dbl>      <dbl>       <dbl>         <dbl>
#>  1 2024-01-01 00:00:00            63         17          12          NA  
#>  2 2024-01-01 01:00:00           102         22          41          NA  
#>  3 2024-01-01 02:00:00           111         95          42          92  
#>  4 2024-01-01 03:00:00            47        129          66          86.7
#>  5 2024-01-01 04:00:00             5         70         102          54.3
#>  6 2024-01-01 05:00:00            24         65         103          25.3
#>  7 2024-01-01 06:00:00            79         25          26          36  
#>  8 2024-01-01 07:00:00            77          8          60          60  
#>  9 2024-01-01 08:00:00           130        108          14          95.3
#> 10 2024-01-01 09:00:00            55         83           5          87.3
#> # ℹ 14 more rows
#> # ℹ 13 more variables: o3_3hr_ppb <dbl>, no2_3hr_ppb <dbl>, level <fct>,
#> #   AQHI <fct>, AQHI_plus <fct>, AQHI_plus_exceeds_AQHI <lgl>,
#> #   AQHI_pm25_ratio <dbl>, AQHI_o3_ratio <dbl>, AQHI_no2_ratio <dbl>,
#> #   colour <chr>, risk <fct>, high_risk_pop_message <chr>,
#> #   general_pop_message <chr>
obs$date |> AQHI(pm25_1hr_ugm3 = obs$pm25, language = "fr")
#> Warning: No non-missing NO2 / O3 data provided. Returning AQHI+ (PM2.5 only) instead of AQHI.
#> # A tibble: 24 × 18
#>    date                pm25_1hr_ugm3 o3_1hr_ppb no2_1hr_ppb pm25_3hr_ugm3
#>    <dttm>                      <dbl>      <dbl>       <dbl>         <dbl>
#>  1 2024-01-01 00:00:00            63         NA          NA            NA
#>  2 2024-01-01 01:00:00           102         NA          NA            NA
#>  3 2024-01-01 02:00:00           111         NA          NA            NA
#>  4 2024-01-01 03:00:00            47         NA          NA            NA
#>  5 2024-01-01 04:00:00             5         NA          NA            NA
#>  6 2024-01-01 05:00:00            24         NA          NA            NA
#>  7 2024-01-01 06:00:00            79         NA          NA            NA
#>  8 2024-01-01 07:00:00            77         NA          NA            NA
#>  9 2024-01-01 08:00:00           130         NA          NA            NA
#> 10 2024-01-01 09:00:00            55         NA          NA            NA
#> # ℹ 14 more rows
#> # ℹ 13 more variables: o3_3hr_ppb <dbl>, no2_3hr_ppb <dbl>, level <fct>,
#> #   AQHI <fct>, AQHI_plus <fct>, AQHI_plus_exceeds_AQHI <lgl>,
#> #   AQHI_pm25_ratio <dbl>, AQHI_o3_ratio <dbl>, AQHI_no2_ratio <dbl>,
#> #   colour <chr>, risk <fct>, high_risk_pop_message <chr>,
#> #   general_pop_message <chr>

# Don't allow AQHI+ override (not recommended)
obs$date |>
  AQHI(
    pm25_1hr_ugm3 = obs$pm25,
    o3_1hr_ppb = obs$o3,
    no2_1hr_ppb = obs$no2,
    allow_aqhi_plus_override = FALSE
  )
#> # A tibble: 24 × 18
#>    date                pm25_1hr_ugm3 o3_1hr_ppb no2_1hr_ppb pm25_3hr_ugm3
#>    <dttm>                      <dbl>      <dbl>       <dbl>         <dbl>
#>  1 2024-01-01 00:00:00            63         17          12          NA  
#>  2 2024-01-01 01:00:00           102         22          41          NA  
#>  3 2024-01-01 02:00:00           111         95          42          92  
#>  4 2024-01-01 03:00:00            47        129          66          86.7
#>  5 2024-01-01 04:00:00             5         70         102          54.3
#>  6 2024-01-01 05:00:00            24         65         103          25.3
#>  7 2024-01-01 06:00:00            79         25          26          36  
#>  8 2024-01-01 07:00:00            77          8          60          60  
#>  9 2024-01-01 08:00:00           130        108          14          95.3
#> 10 2024-01-01 09:00:00            55         83           5          87.3
#> # ℹ 14 more rows
#> # ℹ 13 more variables: o3_3hr_ppb <dbl>, no2_3hr_ppb <dbl>, level <fct>,
#> #   AQHI <fct>, AQHI_plus <fct>, AQHI_plus_exceeds_AQHI <lgl>,
#> #   AQHI_pm25_ratio <dbl>, AQHI_o3_ratio <dbl>, AQHI_no2_ratio <dbl>,
#> #   colour <chr>, risk <fct>, high_risk_pop_message <chr>,
#> #   general_pop_message <chr>