Calculate the Canadian AQHI+ from hourly PM2.5 concentrations

The Canadian AQHI+ is a modification of the Canadian Air Quality Health Index (AQHI). AQHI+ only uses fine particulate matter (PM2.5) instead of the combination of PM2.5, ozone (O3), and nitrogen dioxide (NO2). Unlike the AQHI which uses 3-hourly mean averages, AQHI+ is calculated using hourly mean averages. The AQHI+ overrides the AQHI if it exceeds the AQHI for a particular hour.

AQHI+ categorizes hourly PM2.5 concentrations into 10 equal bins from 0 to 100 μg m-3 (AQHI+ 1-10), and assigns "+" to values greater than 100 μg m-3. The risk categories match the Canadian AQHI and share the same health messaging:

Low: AQHI+ of 1-3, or a concentration of 0-30 μg m-3

Moderate: AQHI+ of 4-6, or a concentration of 30.1-60 μg m-3

High: AQHI+ of 7-10, or a concentration of 60.1-100 μg m-3

Very High: AQHI+ of +, or a concentration above 100 μg m-3

The AQHI+ was originally published by Yao et al (2019): doi:10.17269/s41997-019-00237-w , and has been adopted by all* Canadian provinces/territories as of 2024. (*except Québec where they use the AQI instead of the AQHI/AQHI+).

Quote from Yao et al (2019):

This algorithm is a linear extrapolation based on two anchor points:

the AQHI-Plus value should reach 7 (high risk category), when 1-h PM2.5 concentrations are over 60 ug/m3. This is consistent with [...] the analyses of the likelihood of exceeding the 25 ug/m3 24-h threshold value; and

the AQHI-Plus value should reach 4 (moderate risk category) when the 1-h PM2.5 concentrations are over 30 ug/m3, which is the threshold for odour perception of burning softwood (Kistler et al. 2012: doi:10.1016/j.atmosenv.2012.01.044 ), the major type of forest species in [British Columbia].

[...]

These conclusions suggest that the AQHI is most indicative of the mortality and circulatory risks during periods affected by wildfire smoke. However, the [...] AQHI-Plus is most indicative of the respiratory risks during high-intensity fire periods, particularly for people with asthma. This is consistent with the rationale behind the AQHI formulation and with existing health evidence related to the effects of wildfire smoke.

Usage

AQHI_plus(
  pm25_1hr_ugm3,
  min_allowed_pm25 = 0,
  detailed = TRUE,
  language = "en"
)

Arguments

pm25_1hr_ugm3

A numeric vector of hourly mean PM2.5 concentrations (units = μg m-3). Must be a single value or the same length as dates or a single value (which will be repeated).

min_allowed_pm25

(Optional). A single numeric value indicating the minimum allowed concentration. All values in pm25_1hr_ugm3 less than this will be replaced with NA. Default is 0 μg m-3.

detailed

(Optional). A single logical value indicating if a tibble with levels, risk categories, health messages, etc should be returned. If FALSE only the levels will be returned. Default is TRUE.

language

(Optional). A single character value indicating the language to use for risk levels and health messaging. Must be either "en" (English) or "fr" (French). Not case sensitive. Default is "en".

Value

If detailed is TRUE:

A tibble (data.frame) with columns detailing the AQHI+, risk level, health messages, etc

If detailed is FALSE:

A factor vector of AQHI+ levels with length(pm25_1hr_ugm3) elements

References

Yao et al (2019): doi:10.17269/s41997-019-00237-w

Environment and Climate Change Canada: https://www.canada.ca/en/environment-climate-change/services/air-quality-health-index/about.html

Examples

# Hourly pm2.5 concentrations
pm25 <- sample(1:150, 24)
# Calculate the detailed AQHI+ (tibble with risk levels etc)
AQHI_plus(pm25)
#> # A tibble: 24 × 6
#>    pm25_1hr_ugm3 level colour  risk    high_risk_pop_message general_pop_message
#>            <int> <fct> <chr>   <fct>   <chr>                 <chr>              
#>  1           109 +     #650205 Very H… Avoid strenuous acti… Reduce or reschedu…
#>  2            25 3     #0D6797 Low     Enjoy your usual act… Ideal air quality …
#>  3            53 6     #FE9A3F Modera… Consider reducing or… No need to modify …
#>  4            45 5     #FFCC2E Modera… Consider reducing or… No need to modify …
#>  5             4 1     #21C6F5 Low     Enjoy your usual act… Ideal air quality …
#>  6            41 5     #FFCC2E Modera… Consider reducing or… No need to modify …
#>  7            82 9     #FF0101 High    Reduce or reschedule… Consider reducing …
#>  8            94 10    #CB0713 High    Reduce or reschedule… Consider reducing …
#>  9            51 6     #FE9A3F Modera… Consider reducing or… No need to modify …
#> 10           128 +     #650205 Very H… Avoid strenuous acti… Reduce or reschedu…
#> # ℹ 14 more rows
# Or just the AQHI+ (returned as a factor)
AQHI_plus(pm25, detailed = FALSE)
#>  [1] +  3  6  5  1  5  9  10 6  +  2  4  1  1  2  +  5  4  +  2  5  9  6  3 
#> Levels: 1 2 3 4 5 6 7 8 9 10 +

# Hourly pm2.5 concentrations (with negative values)
pm25 <- c(-2, -0.1, sample(1:150, 22))
# Calculate the AQHI+ for each hour, except for hours where pm2.5 is < -0.5
AQHI_plus(pm25, min_allowed_pm25 = -0.5)
#> # A tibble: 24 × 6
#>    pm25_1hr_ugm3 level colour  risk    high_risk_pop_message general_pop_message
#>            <dbl> <fct> <chr>   <fct>   <chr>                 <chr>              
#>  1          NA   NA    #bbbbbb NA      NA                    NA                 
#>  2          -0.1 1     #21C6F5 Low     Enjoy your usual act… Ideal air quality …
#>  3          74   8     #FF3B3B High    Reduce or reschedule… Consider reducing …
#>  4         102   +     #650205 Very H… Avoid strenuous acti… Reduce or reschedu…
#>  5         142   +     #650205 Very H… Avoid strenuous acti… Reduce or reschedu…
#>  6          66   7     #FD6769 High    Reduce or reschedule… Consider reducing …
#>  7          37   4     #FFFD37 Modera… Consider reducing or… No need to modify …
#>  8         108   +     #650205 Very H… Avoid strenuous acti… Reduce or reschedu…
#>  9         132   +     #650205 Very H… Avoid strenuous acti… Reduce or reschedu…
#> 10          83   9     #FF0101 High    Reduce or reschedule… Consider reducing …
#> # ℹ 14 more rows